Added buffered channels to verify + docs

+++ coconut/concurrency/jobworker/job_worker.go
@@ -1,3 +1,20 @@
+// job_worker.go - Worker for job queue tasks.
+// Copyright (C) 2018  Jedrzej Stuczynski.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as
+// published by the Free Software Foundation, either version 3 of the
+// License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+// Package jobworker implements worker for performing tasks defined by jobpacket that are in the queue.
 package jobworker
 
 import (
@@ -5,10 +22,11 @@ import (
 	"github.com/jstuczyn/CoconutGo/coconut/concurrency/worker"
 )
 
+// Worker is an instance of jobWorker.
 type Worker struct {
 	worker.Worker
 
-	id       int
+	id       uint64
 	jobQueue <-chan interface{}
 }
 
@@ -30,7 +48,8 @@ func (w *Worker) worker() {
 	}
 }
 
-func New(jobQueue <-chan interface{}, id int) *Worker {
+// New creates new instance of a jobWorker.
+func New(jobQueue <-chan interface{}, id uint64) *Worker {
 	w := &Worker{
 		jobQueue: jobQueue,
 		id:       id,
